Make two-factor login more JSON friendly
