+++ flask_security/views.py
@@ -14,6 +14,7 @@ from flask import current_app, redirect, request, jsonify, \
 from flask_login import current_user
 from werkzeug.datastructures import MultiDict
 from werkzeug.local import LocalProxy
+import pyqrcode
 
 from .changeable import change_user_password
 from .confirmable import confirm_email_token_status, confirm_user, \
@@ -27,8 +28,8 @@ from .utils import config_value, do_flash, get_url, get_post_login_redirect, \
     get_post_register_redirect, get_message, login_user, logout_user, \
     url_for_security as url_for, slash_url_suffix, send_mail,\
     get_post_logout_redirect
-from .twofactor import send_security_token, generate_totp, generate_qrcode, \
-    complete_two_factor_process
+from .twofactor import send_security_token, generate_totp, \
+    complete_two_factor_process, get_totp_uri
 
 # Convenient references
 _security = LocalProxy(lambda: current_app.extensions['security'])
@@ -132,6 +133,7 @@ def register():
                 redirect_url = get_post_register_redirect()
 
             return redirect(redirect_url)
+
         return _render_json(form, include_auth_token=True)
 
     if request.is_json:
@@ -353,23 +355,26 @@ def two_factor_login():
         user = form.user
         session['email'] = user.email
         # if user's two-factor properties are not configured
-        if not request.is_json:
-            if user.two_factor_primary_method is None or user.totp_secret is None:
-                session['has_two_factor'] = False
+
+        if user.two_factor_primary_method is None or\
+                user.totp_secret is None:
+            session['has_two_factor'] = False
+            if not request.is_json:
                 return redirect(url_for('two_factor_setup_function'))
-            # if user's two-factor properties are configured
-            else:
-                session['has_two_factor'] = True
-                session['primary_method'] = user.two_factor_primary_method
-                session['totp_secret'] = user.totp_secret
-                send_security_token(user=user,
-                                    method=user.two_factor_primary_method,
-                                    totp_secret=user.totp_secret)
+
+        # if user's two-factor properties are configured
+        else:
+            session['has_two_factor'] = True
+            session['primary_method'] = user.two_factor_primary_method
+            session['totp_secret'] = user.totp_secret
+            send_security_token(user=user,
+                                method=user.two_factor_primary_method,
+                                totp_secret=user.totp_secret)
+            if not request.is_json:
                 return redirect(url_for('two_factor_token_validation'))
 
     if request.is_json:
-        form.user = current_user
-        return _render_json(form)
+        return _render_json(form, include_user=False)
 
     return _security.render_template(config_value('LOGIN_USER_TEMPLATE'),
                                      login_user_form=form,
@@ -381,22 +386,6 @@ def two_factor_setup_function():
 
     # user's email&password not approved or we are
     # logged in and didn't validate password
-    if not request.is_json:
-        if 'password_confirmed' not in session:
-            if 'email' not in session or 'has_two_factor' not in session:
-                do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
-                return redirect(get_url(_security.login_url))
-
-            # user's email&password approved and
-            # two-factor properties were configured before
-            if session['has_two_factor'] is True:
-                do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
-                return redirect(url_for('two_factor_token_validation'))
-
-            user = _datastore.find_user(email=session['email'])
-        else:
-            user = current_user
-
     form_class = _security.two_factor_setup_form
 
     if request.is_json:
@@ -404,8 +393,34 @@ def two_factor_setup_function():
     else:
         form = form_class()
 
+    if 'password_confirmed' not in session:
+
+        if 'email' not in session or 'has_two_factor' not in session:
+            if not request.is_json:
+                do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
+                return redirect(get_url(_security.login_url))
+            else:
+                m, c = get_message('TWO_FACTOR_PERMISSION_DENIED')
+                form._errors = m
+                return _render_json(form)
+
+        # user's email&password approved and
+        # two-factor properties were configured before
+        if session['has_two_factor'] is True:
+            if not request.is_json:
+                do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
+                return redirect(url_for('two_factor_token_validation'))
+            else:
+                m, c = get_message('TWO_FACTOR_PERMISSION_DENIED')
+                form._errors = m
+                return _render_json(form)
+
+        user = _datastore.find_user(email=session['email'])
+    else:
+        user = current_user
+
     if form.validate_on_submit():
-        # totp and primarty_method are added to
+        # totp and primary_method are added to
         # session to flag the user's temporary choice
         session['totp_secret'] = generate_totp()
         session['primary_method'] = form['setup'].data
@@ -427,23 +442,19 @@ def two_factor_setup_function():
     if request.is_json:
         return _render_json(form, include_user=False)
 
+    code_form = _security.two_factor_verify_code_form()
+    return _security.render_template(
+        config_value('TWO_FACTOR_CHOOSE_METHOD_TEMPLATE'),
+        two_factor_setup_form=form,
+        two_factor_verify_code_form=code_form,
+        choices=config_value(
+            'TWO_FACTOR_ENABLED_METHODS'),
+        **_ctx('two_factor_setup_function'))
+
 
 def two_factor_token_validation():
     """View function for two-factor token validation during login process"""
     # if we are in login process and not changing current two-factor method
-    if not request.is_json:
-        if 'password_confirmed' not in session:
-            # user's email&password not approved or we are logged in
-            # and didn't validate password
-            if 'has_two_factor' not in session:
-                do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
-                return redirect(get_url(_security.login_url))
-
-        # make sure user has or has chosen a two-factor
-        # method before we try to validate
-        if 'totp_secret' not in session or 'primary_method' not in session:
-            do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
-            return redirect(url_for('two_factor_setup_function'))
 
     form_class = _security.two_factor_verify_code_form
 
@@ -452,17 +463,42 @@ def two_factor_token_validation():
     else:
         form = form_class()
 
+    if 'password_confirmed' not in session:
+        # user's email&password not approved or we are logged in
+        # and didn't validate password
+        if 'has_two_factor' not in session:
+            if not request.is_json:
+                do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
+                return redirect(get_url(_security.login_url))
+            else:
+                m, c = get_message('TWO_FACTOR_PERMISSION_DENIED')
+                form._errors = m
+                return _render_json(form)
+        # make sure user has or has chosen a two-factor
+        # method before we try to validate
+        if 'totp_secret' not in session or 'primary_method' not in session:
+            if not request.is_json:
+                do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
+                return redirect(url_for('two_factor_setup_function'))
+            else:
+                m, c = get_message('TWO_FACTOR_PERMISSION_DENIED')
+                form._errors = m
+                return _render_json(form)
+
     if form.validate_on_submit():
         complete_two_factor_process(form.user)
         after_this_request(_commit)
-        return redirect(get_post_login_redirect())
+        if not request.is_json:
+            return redirect(get_post_login_redirect())
 
     if request.is_json:
-        return _render_json(form, include_user=False)
+        form.user = current_user
+        return _render_json(form)
 
     # if we were trying to validate a new method
     if 'password_confirmed' in session or session['has_two_factor'] is False:
         setup_form = _security.two_factor_setup_form()
+
         return _security.render_template(
             config_value('TWO_FACTOR_CHOOSE_METHOD_TEMPLATE'),
             two_factor_setup_form=setup_form,
@@ -470,9 +506,12 @@ def two_factor_token_validation():
             choices=config_value(
                 'TWO_FACTOR_ENABLED_METHODS'),
             **_ctx('two_factor_setup_function'))
+
     # if we were trying to validate an existing method
     else:
+
         rescue_form = _security.two_factor_rescue_form()
+
         return _security.render_template(
             config_value('TWO_FACTOR_VERIFY_CODE_TEMPLATE'),
             two_factor_rescue_form=rescue_form,
@@ -486,16 +525,6 @@ def two_factor_rescue_function():
     """ Function that handles a situation where user can't
     enter his two-factor validation code"""
     # user's email&password yet to be approved
-    if not request.is_json:
-        if 'email' not in session:
-            do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
-            return redirect(get_url(_security.login_url))
-
-        # user's email&password approved and two-factor properties
-        # were not configured
-        if 'totp_secret' not in session or 'primary_method' not in session:
-            do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
-            return redirect(get_url(_security.login_url))
 
     form_class = _security.two_factor_rescue_form
 
@@ -503,6 +532,26 @@ def two_factor_rescue_function():
         form = form_class(MultiDict(request.get_json()))
     else:
         form = form_class()
+
+    if 'email' not in session:
+        if not request.is_json:
+            do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
+            return redirect(get_url(_security.login_url))
+        else:
+            m, c = get_message('TWO_FACTOR_PERMISSION_DENIED')
+            form._errors = m
+            return _render_json(form, include_user=False)
+    # user's email&password approved and two-factor properties
+    # were not configured
+    if 'totp_secret' not in session or 'primary_method' not in session:
+        if not request.is_json:
+            do_flash(*get_message('TWO_FACTOR_PERMISSION_DENIED'))
+            return redirect(get_url(_security.login_url))
+
+        else:
+            m, c = get_message('TWO_FACTOR_PERMISSION_DENIED')
+            form._errors = m
+            return _render_json(form, include_user=False)
 
     problem = None
     if form.validate_on_submit():
@@ -551,6 +600,11 @@ def two_factor_password_confirmation():
             do_flash(get_message('TWO_FACTOR_PASSWORD_CONFIRMATION_DONE'))
             return redirect(url_for('two_factor_setup_function'))
 
+        else:
+            m, c = get_message('TWO_FACTOR_PASSWORD_CONFIRMATION_DONE')
+            form._errors = m
+            return _render_json(form)
+
     if request.is_json:
         form.user = current_user
         return _render_json(form)
@@ -564,6 +618,35 @@ def two_factor_password_confirmation():
 
 def two_factor_qrcode():
     return generate_qrcode()
+
+
+def generate_qrcode():
+    if 'google_authenticator' not in\
+            config_value('TWO_FACTOR_ENABLED_METHODS'):
+        return abort(404)
+    if 'primary_method' not in session or\
+        session['primary_method'] != 'google_authenticator' \
+            or 'totp_secret' not in session:
+        return abort(404)
+
+    if 'email' in session:
+        email = session['email']
+    elif 'password_confirmed' in session:
+        email = current_user.email
+    else:
+        return abort(404)
+
+    name = email.split('@')[0]
+    totp = session['totp_secret']
+    url = pyqrcode.create(get_totp_uri(name, totp))
+    from io import BytesIO
+    stream = BytesIO()
+    url.svg(stream, scale=3)
+    return stream.getvalue(), 200, {
+        'Content-Type': 'image/svg+xml',
+        'Cache-Control': 'no-cache, no-store, must-revalidate',
+        'Pragma': 'no-cache',
+        'Expires': '0'}
 
 
 def create_blueprint(state, import_name):
